backend/core/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
build==1.3.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.2
django-cors-headers==4.4.0
django-filter==24.2
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
factory_boy==3.3.3
Faker==37.5.3
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
openapi==2.0.0
packaging==25.0
pillow==11.3.0
pip-tools==7.5.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.7.0
